+++ test/cleanup.test.js
@@ -1,16 +1,4 @@
-/**
- * Copyright(c) cnpm and other contributors.
- * MIT Licensed
- *
- * Authors:
- *   dead_horse <dead_horse@qq.com>
- */
-
 'use strict';
-
-/**
- * Module dependencies.
- */
 
 const assert = require('assert');
 const path = require('path');
@@ -32,7 +20,7 @@ describe('test/cleanup.test.js', function() {
   });
   afterEach(cleanup);
 
-  it('should remove donefile when install failed', function*() {
+  it('should remove donefile when install failed', function* () {
     let throwError = false;
     try {
       yield npminstall({
@@ -46,10 +34,10 @@ describe('test/cleanup.test.js', function() {
     }
     assert.equal(throwError, true);
 
-    let exists = yield fs.exists(path.join(tmp, 'node_modules/.npminstall/install-error/1.0.1/install-error'));
+    let exists = yield fs.exists(path.join(tmp, 'node_modules/.install-error@1.0.1/.npminstall.done'));
     assert.equal(exists, false);
     const dirs = yield fs.readdir(path.join(tmp, 'node_modules'));
-    assert.deepEqual(dirs, ['.npminstall']);
+    assert.deepEqual(dirs, [ '.install-error@1.0.1' ]);
 
     // install again will try to download
     throwError = false;
@@ -64,7 +52,38 @@ describe('test/cleanup.test.js', function() {
       throwError = true;
     }
     assert.equal(throwError, true);
-    exists = yield fs.exists(path.join(tmp, 'node_modules/.npminstall/install-error/1.0.1/install-error'));
+    exists = yield fs.exists(path.join(tmp, 'node_modules/.install-error@1.0.1/.npminstall.done'));
+    assert.equal(exists, false);
+  });
+
+  it('should remove donefile when execute postinstall script failed', function* () {
+    let throwError = false;
+    const pkgs = [{ version: '../postinstall-error', type: 'local' }];
+    try {
+      yield npminstall({
+        root: tmp,
+        pkgs,
+      });
+    } catch (err) {
+      throwError = true;
+    }
+    assert.equal(throwError, true);
+
+    let exists = yield fs.exists(path.join(tmp, 'node_modules/.postinstall-error@1.0.0/.npminstall.done'));
+    assert.equal(exists, false);
+
+    // install again will try to download
+    throwError = false;
+    try {
+      yield npminstall({
+        root: tmp,
+        pkgs,
+      });
+    } catch (err) {
+      throwError = true;
+    }
+    assert.equal(throwError, true);
+    exists = yield fs.exists(path.join(tmp, 'node_modules/.postinstall-error@1.0.0/.npminstall.done'));
     assert.equal(exists, false);
   });
 });