+++ test/installLocal.test.js
@@ -1,23 +1,10 @@
-/**
- * Copyright(c) cnpm and other contributors.
- * MIT Licensed
- *
- * Authors:
- *   dead_horse <dead_horse@qq.com>
- */
-
 'use strict';
-
-/**
- * Module dependencies.
- */
 
 const assert = require('assert');
 const path = require('path');
 const rimraf = require('rimraf');
 const readJSON = require('../lib/utils').readJSON;
 const npminstall = require('./npminstall');
-const fs = require('mz/fs');
 
 describe('test/installLocal.test.js', function() {
   const root = path.join(__dirname, 'fixtures', 'local');
@@ -29,9 +16,9 @@ describe('test/installLocal.test.js', function() {
   beforeEach(cleanup);
   afterEach(cleanup);
 
-  it('should install local folder ok', function*() {
+  it('should install local folder ok', function* () {
     yield npminstall({
-      root: root,
+      root,
       pkgs: [
         { name: null, version: 'file:pkg' },
       ],
@@ -40,9 +27,9 @@ describe('test/installLocal.test.js', function() {
     assert.equal(pkg.name, 'pkg');
   });
 
-  it('should install local folder with relative path ok', function*() {
+  it('should install local folder with relative path ok', function* () {
     yield npminstall({
-      root: root,
+      root,
       pkgs: [
         { name: null, version: './pkg' },
       ],
@@ -51,12 +38,12 @@ describe('test/installLocal.test.js', function() {
     assert.equal(pkg.name, 'pkg');
   });
 
-  it('should install local link folder ok', function*() {
+  it('should install local link folder ok', function* () {
     if (process.platform === 'win32') {
       return;
     }
     yield npminstall({
-      root: root,
+      root,
       pkgs: [
         { name: null, version: 'file:pkg-link' },
       ],
@@ -65,9 +52,9 @@ describe('test/installLocal.test.js', function() {
     assert.equal(pkg.name, 'pkg');
   });
 
-  it('should install local gzip tarball ok', function*() {
+  it('should install local gzip tarball ok', function* () {
     yield npminstall({
-      root: root,
+      root,
       pkgs: [
         { name: null, version: 'file:sequelize.tgz' },
       ],
@@ -77,12 +64,12 @@ describe('test/installLocal.test.js', function() {
     assert.equal(pkg.name, 'sequelize');
   });
 
-  it('should install local link gzip tarball ok', function*() {
+  it('should install local link gzip tarball ok', function* () {
     if (process.platform === 'win32') {
       return;
     }
     yield npminstall({
-      root: root,
+      root,
       pkgs: [
         { name: null, version: 'file:sequelize-link.tgz' },
       ],
@@ -92,9 +79,9 @@ describe('test/installLocal.test.js', function() {
     assert.equal(pkg.name, 'sequelize');
   });
 
-  it('should install local naked tarball ok', function*() {
+  it('should install local naked tarball ok', function* () {
     yield npminstall({
-      root: root,
+      root,
       pkgs: [
         { name: null, version: 'file:pkg.tar' },
       ],
@@ -103,10 +90,10 @@ describe('test/installLocal.test.js', function() {
     assert.equal(pkg.name, 'pkg');
   });
 
-  it('should install local folder without package.json error', function*() {
+  it('should install local folder without package.json error', function* () {
     try {
       yield npminstall({
-        root: root,
+        root,
         pkgs: [
           { name: null, version: 'file:not-pkg' },
         ],
@@ -117,10 +104,10 @@ describe('test/installLocal.test.js', function() {
     }
   });
 
-  it('should install local tarball without package.json error', function*() {
+  it('should install local tarball without package.json error', function* () {
     try {
       yield npminstall({
-        root: root,
+        root,
         pkgs: [
           { name: null, version: 'file:not-pkg.tar' },
         ],
@@ -131,10 +118,10 @@ describe('test/installLocal.test.js', function() {
     }
   });
 
-  it('should install local folder without package name error', function*() {
+  it('should install local folder without package name error', function* () {
     try {
       yield npminstall({
-        root: root,
+        root,
         pkgs: [
           { name: null, version: 'file:pkg-without-name' },
         ],
@@ -145,10 +132,10 @@ describe('test/installLocal.test.js', function() {
     }
   });
 
-  it('should install local tarball without package name error', function*() {
+  it('should install local tarball without package name error', function* () {
     try {
       yield npminstall({
-        root: root,
+        root,
         pkgs: [
           { name: null, version: 'file:pkg-without-name.tgz' },
         ],
@@ -160,9 +147,9 @@ describe('test/installLocal.test.js', function() {
   });
 
   if (process.platform !== 'win32') {
-    it('should install the same tarball ok', function*() {
+    it('should install the same tarball ok', function* () {
       yield npminstall({
-        root: root,
+        root,
         pkgs: [
           { name: null, version: 'file:pkg.tar' },
         ],
@@ -170,7 +157,7 @@ describe('test/installLocal.test.js', function() {
       let pkg = yield readJSON(path.join(root, 'node_modules/pkg/package.json'));
       assert.equal(pkg.name, 'pkg');
       yield npminstall({
-        root: root,
+        root,
         pkgs: [
           { name: null, version: 'file:pkg.tar' },
           // { name: null, version: 'file:pkg.tar' },
@@ -178,13 +165,11 @@ describe('test/installLocal.test.js', function() {
       });
       pkg = yield readJSON(path.join(root, 'node_modules/pkg/package.json'));
       assert.equal(pkg.name, 'pkg');
-      const versions = yield fs.readdir(path.join(root, 'node_modules/.npminstall/pkg'));
-      assert.equal(versions.length, 1);
     });
 
-    it('should install the same local folder ok', function*() {
+    it('should install the same local folder ok', function* () {
       yield npminstall({
-        root: root,
+        root,
         pkgs: [
           { name: null, version: 'file:pkg' },
         ],
@@ -192,7 +177,7 @@ describe('test/installLocal.test.js', function() {
       let pkg = yield readJSON(path.join(root, 'node_modules/pkg/package.json'));
       assert.equal(pkg.name, 'pkg');
       yield npminstall({
-        root: root,
+        root,
         pkgs: [
           { name: null, version: 'file:pkg' },
           { name: null, version: 'file:pkg' },
@@ -200,8 +185,6 @@ describe('test/installLocal.test.js', function() {
       });
       pkg = yield readJSON(path.join(root, 'node_modules/pkg/package.json'));
       assert.equal(pkg.name, 'pkg');
-      const versions = yield fs.readdir(path.join(root, 'node_modules/.npminstall/pkg'));
-      assert.equal(versions.length, 1);
     });
   }
 });